Fix instructions for running knex in non-dev environments
The current instructions specify usage of the knex CLI's --env parameter, but this
assumes that the knexfile itself has configuration objects for multiple environments.
(http://knexjs.org/#knexfile)

However, the current knexfile.js leverages the npm 'config' module, which presents a
single config to Knex based on the NODE_ENV variable.  The --env parameter has no effect
on Knex with this knexfile.js.